requirements.txt
flask
flask_socketio
eventlet
python-dotenv